Add global vars file
a few changes that went on in addition to global_vars.py
disc_pretraining is now called in toy trials, but only iterates through once. This is to allow all the code to be tested in toy trials.
Move process_opts out of pg_gan and summary_stats and into util. Consolidates nearly identical code, with a toggle for changes corresponding to summary_stats. This function also has a simplified way to build the population sizes array.
In real_data_random, when reading data sometimes the chrom value is a byte, and sometimes it's stored as a string. Added a line to make sure that the chrom number is always printed as a string and not a  byte.